Added initial converter functionality

+++ app/src/main/java/az/siftoshka/junkyconverter/utils/Converter.kt
@@ -7,5 +7,5 @@ import java.text.DecimalFormat
  */
 fun String.moneyFormat(): String {
     if (this == "0") return this
-    return DecimalFormat("##,###.00").format(this.toInt())
+    return DecimalFormat("##,###.00").format(this.toFloat())
 }
